% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cik_lookup.R
\name{cik_lookup}
\alias{cik_lookup}
\title{Look up a company's CIK.}
\usage{
cik_lookup(ticker, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A data frame of the chosen company's CIK.
}
\description{
Enter function parameters in parentheses.
}
