% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_entity_summary.R
\name{doc_entity_summary}
\alias{doc_entity_summary}
\title{Display a company's documentation and entity information.}
\usage{
doc_entity_summary(ticker, form_type, filing_order, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{form_type}{"10-K" or "10-Q".}

\item{filing_order}{"0" for most recent statement, "1" for prior statement, "2" for next oldest statement, etc.}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
Data frame of documentation and entity information from this company's filing.
}
\description{
Enter function parameters in parentheses.
}
