% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnings_per_share_data.R
\name{earnings_per_share_data}
\alias{earnings_per_share_data}
\title{Create a data frame of a company's earnings per share.}
\usage{
earnings_per_share_data(ticker, form_type, filing_order, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{form_type}{"10-K" or "10-Q".}

\item{filing_order}{"0" for most recent statement, "1" for prior statement, "2" for next oldest statement, etc.}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A data frame of the chosen company's earnings per share.
}
\description{
Enter function parameters in parentheses.
}
