% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_ratios.R
\name{fin_ratios}
\alias{fin_ratios}
\title{Display a company's financial ratios.}
\usage{
fin_ratios(ticker, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A list of the chosen company's financial ratios.
}
\description{
Enter function parameters in parentheses.
}
