% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_stock_prices.R
\name{historical_stock_prices}
\alias{historical_stock_prices}
\title{Create a data frame of a company's historical stock prices.}
\usage{
historical_stock_prices(ticker, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A data frame of the chosen company's historical stock prices.
}
\description{
Enter function parameters in parentheses.
}
