% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liabilities_data.R
\name{liabilities_data}
\alias{liabilities_data}
\title{Create a data frame of a company's recent liabilities.}
\usage{
liabilities_data(ticker, form_type, filing_order, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{form_type}{"10-K" or "10-Q".}

\item{filing_order}{"0" for most recent statement, "1" for prior statement, "2" for next oldest statement, etc.}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A data frame of the chosen company's recent liabilities from past transactions or events.
}
\description{
Enter function parameters in parentheses.
}
