% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_full.R
\name{operations_full}
\alias{operations_full}
\title{Display all statement of operations content from a company's filing.}
\usage{
operations_full(ticker, form_type, filing_order, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{form_type}{"10-K" or "10-Q".}

\item{filing_order}{"0" for most recent statement, "1" for prior statement, "2" for next oldest statement, etc.}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
All statement of operations content from this company's filing.
}
\description{
Enter function parameters in parentheses.
}
