% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_quote.R
\name{stock_quote}
\alias{stock_quote}
\title{Create a data frame of a company's stock quote information.}
\usage{
stock_quote(ticker, subscription_key)
}
\arguments{
\item{ticker}{Company stock ticker, for example "AAPL".}

\item{subscription_key}{Your Last10k.com Primary key, in parentheses, shown under Your Subscriptions in your profile page.}
}
\value{
A data frame of the chosen company's stock quote.
}
\description{
Enter function parameters in parentheses.
}
